requirements.txt
pandas
numpy
requests
scikit-learn
joblib
whisper
opencv-python
transformers 
torch 
pillow
sentence-transformers
